@@ -1,12 +1,14 @@
 name: flutter_background_geolocation
 description: The most sophisticated background location tracking & geofencing module with battery-conscious motion-detection intelligence for iOS and Android.
-version: 2.0.0
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/transistorsoft/flutter_background_geolocation
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.2.4
+  meta: ^1.3.0
+  collection: ^1.15.0
+
   background_fetch: ^1.0.0-nullsafety.1
 
 flutter:
@@ -19,6 +21,6 @@ flutter:
         pluginClass: FLTBackgroundGeolocationPlugin
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
