@@ -5,16 +5,16 @@ part of flt_background_geolocation;
 class AuthorizationEvent {
   /// `true` when an authorization request to [Authorization.refreshUrl] was successful.
   ///
-  bool success;
+  late bool success;
 
   /// When [success] `false`, this is the error message from [Authorization.refreshUrl].
-  String error;
+  String? error;
 
   /// when [success] `true`, this is the data returned from [Authorization.refreshUrl].
-  Map response;
+  Map? response;
 
   AuthorizationEvent(Map data) {
-    success = _ensureBool(data['success']);
+    success = _ensureBool(data['success'])!;
     error = data['error'];
     response = data['response'];
   }
