@@ -444,7 +444,7 @@ public class BackgroundGeolocationModule  implements MethodChannel.MethodCallHan
                 result.success(isMoving);
             }
             @Override public void onFailure(String error) {
-                result.error(error, null, null);
+                result.error("0", error, null);
             }
         });
     }
